requirements.txt
asgiref==3.8.1
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
drf-yasg==1.21.9
googlemaps==4.10.0
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
reportlab==4.3.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
